fix: sql

+++ src/form/controllers/raw_form_id.rs
@@ -3,5 +3,5 @@ use serde::Deserialize;
 
 #[derive(Deserialize, Getters)]
 pub struct RawFormId {
-    form_id: u8,
+    form_id: i32,
 }
